pyproject.toml
[project]
name = "enterprise-ai-code-reviewer"
version = "1.0.0"
description = "Enterprise-grade AI Code Review Agent with CI/CD integration"
readme = "README.md"
requires-python = "==3.12.12"
dependencies = [
    "python-gitlab==4.4.0",
    "openai==1.12.0",
    "requests",
    "httpx",
    "langchain==0.1.0",
    "langchain-community==0.0.10",
    "PyYAML",
    "langchain-openai==0.0.2"
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov"
]
